requirements.txt
gunicorn
jinja2
psycopg2-binary
#cryptography
http-ece
requests
six
py-vapid
